documented the various rtisan functions and some cleanups.
In my own implementation, I made the Nb calculation a bit more
functional: I separated the side-effects of writing the metadata from
the actual calculations.  This should prove useful when/if I eventually
pull out the Nb calculation performed by rtisan.

With that goal in mind, I started documenting the rtisan code in its
current form.  Finally, I implemented an initial vignette in which I
used the rtisan functions with the tags I extracted in read_csv_tags().
Thus, with some minor changes we should be able to start pulling tags
from the qiime otus and/or read_idx.